+++ rtic-monotonics/CHANGELOG.md
@@ -7,6 +7,10 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ## Unreleased
 
+### Fixed
+
+- **Soundness fix:** Monotonics did not wait long enough in `Duration` based delays.
+
 ## v1.3.0 - 2023-11-08
 
 ### Added